Safari-611.1.10/LayoutTests/js/dom/document-all-typeof-is-function-fold-expected.txt

typeof document.all is never 'function'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testTypeofIsFunction() is false
PASS testTypeofIsNotFunction() is true
PASS successfullyParsed is true

TEST COMPLETE
